Refactor prometheus config
