+++ examples/simple_client/notif/development.ini
@@ -19,8 +19,8 @@ blacksmith.client.consul_sd_config =
 blacksmith.client.middlewares =
     prometheus
 
-blacksmith.client.middleware.prometheus.buckets =
-    0.1 0.2 0.4 1.6 3.2 6.4 12.8
+blacksmith.client.middleware.prometheus =
+    buckets 0.1 0.2 0.4 1.6 3.2 6.4 12.8
 
 ###
 # wsgi server configuration
