+++ tests/unittests/test_binding.py
@@ -19,7 +19,6 @@ from pyramid.interfaces import IRequestExtensions
 from pyramid_blacksmith.binding import (
     BlacksmithClientSettingsBuilder,
     PyramidBlacksmith,
-    list_to_dict,
 )
 
 here = Path(__file__).parent.parent.parent
@@ -53,40 +52,6 @@ def test_includeme(config):
     }
     assert registry.clients["api"]["dummy"].collection is None
     assert registry.clients["api"]["dummy"].resource.path == "/dummies/{name}"
-
-
-@pytest.mark.parametrize(
-    "params",
-    [
-        {
-            "settings": {
-                "key": [
-                    "api/v1 http://api.v1",
-                    "smtp smtp://host/",
-                ]
-            },
-            "expected": {
-                "api/v1": "http://api.v1",
-                "smtp": "smtp://host/",
-            },
-        },
-        {
-            "settings": {
-                "key": """
-                    api/v1 http://api.v1
-                    smtp   smtp://host/
-                """
-            },
-            "expected": {
-                "api/v1": "http://api.v1",
-                "smtp": "smtp://host/",
-            },
-        },
-    ],
-)
-def test_list_to_dict(params):
-    dict_ = list_to_dict(params["settings"], "key")
-    assert dict_ == params["expected"]
 
 
 @pytest.mark.parametrize(
